move components into containers and modals folders
