+++ src/client/App.tsx
@@ -1,4 +1,4 @@
-import MainContainer from './components/MainContainer';
+import MainContainer from './components/Containers/MainContainer';
 import React from 'react';
 
 export default function App() {
