+++ src/client/components/Containers/Preview.tsx
@@ -1,8 +1,8 @@
-import CodePreview from './CodePreview';
-import FlowTree from './FlowTree';
-import { Item } from './../../types';
+import CodePreview from '../CodePreview';
+import FlowTree from '../FlowTree';
+import { Item } from '../../../types';
 import React, { useEffect, useState, useContext } from 'react';
-import { PlaygroundContext } from './Playground';
+import { PlaygroundContext } from '../Playground';
 
 export default function Preview(props: { tags: Item[]; currComp: Item }) {
   const { comps } = useContext(PlaygroundContext);
