--- src/client/components/Display.tsx
@@ -1,7 +1,0 @@
-export default function Display(){
-    return(
-        <div className="border-yellow-600 border-2">
-            Display
-        </div>
-    )
-};
